api/requirements.txt
Flask==1.0.2
fpdf==1.7.2
image==1.5.24
Pillow==5.1.0
